add oncoprint track data for patient view page
