+++ src/pages/resultsView/ResultsViewPageStoreUtils.spec.ts
@@ -10,10 +10,8 @@ import {
     CancerStudy,
 } from 'cbioportal-ts-api-client';
 import {
-    annotateMolecularDatum,
     annotateMutationPutativeDriver,
     computeCustomDriverAnnotationReport,
-    computeGenePanelInformation,
     filterAndAnnotateMutations,
     filterSubQueryData,
     getOncoKbOncogenic,
@@ -1121,335 +1119,6 @@ describe('ResultsViewPageStoreUtils', () => {
         });
     });
 
-    describe('annotateMolecularDatum', () => {
-        it('annotates single element correctly in case of Likely Oncogenic', () => {
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({
-                            oncogenic: 'Likely Oncogenic',
-                        } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: 'Likely Oncogenic',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-        });
-        it('annotates single element correctly in case of Predicted Oncogenic', () => {
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({
-                            oncogenic: 'Predicted Oncogenic',
-                        } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: 'Predicted Oncogenic',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-        });
-        it('annotates single element correctly in case of Oncogenic', () => {
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Oncogenic' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: 'Oncogenic',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-        });
-        it('annotates single element correctly in case of Likely Neutral, Inconclusive, Unknown, asdfasd, undefined, empty', () => {
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Likely Neutral' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Inconclusive' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Unknown' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'asdfasdf' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: undefined } as any),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: '' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'COPY_NUMBER_ALTERATION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-        });
-        it('annotates non-copy number data with empty string', () => {
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Oncogenic' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'MUTATION_EXTENDED',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Oncogenic' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'MRNA_EXPRESSION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Oncogenic' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'PROTEIN_LEVEL',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-            assert.deepEqual(
-                annotateMolecularDatum(
-                    {
-                        value: 0,
-                        molecularProfileId: 'profile',
-                        entrezGeneId: 9,
-                    } as NumericGeneMolecularData,
-                    (d: NumericGeneMolecularData) =>
-                        ({ oncogenic: 'Oncogenic' } as IndicatorQueryResp),
-                    {
-                        profile: {
-                            molecularAlterationType: 'FUSION',
-                        } as MolecularProfile,
-                    },
-                    { 9: { hugoGeneSymbol: 'genesymbol' } as Gene }
-                ),
-                {
-                    value: 0,
-                    molecularProfileId: 'profile',
-                    oncoKbOncogenic: '',
-                    hugoGeneSymbol: 'genesymbol',
-                    entrezGeneId: 9,
-                } as any
-            );
-        });
-    });
-
     describe('getDefaultSelectedStudiesForExpressionTab', () => {
         it('recognizes pub tcga study', () => {
             const studyId = 'blca_tcga_pub';
@@ -1470,363 +1139,6 @@ describe('ResultsViewPageStoreUtils', () => {
             assert.isFalse(isTCGAPubStudy(studyId));
             assert.isFalse(isTCGAProvStudy(studyId));
             assert.isTrue(isPanCanStudy(studyId));
-        });
-    });
-
-    describe('computeGenePanelInformation', () => {
-        const genes: Gene[] = [];
-        const samples: Sample[] = [];
-        const patients: Patient[] = [];
-        let genePanelDatum1: any,
-            genePanelDatum2: any,
-            wxsDatum1: any,
-            nsDatum1: any,
-            nsDatum2: any;
-        let genePanels: any[] = [];
-        before(() => {
-            genes.push({
-                entrezGeneId: 0,
-                hugoGeneSymbol: 'GENE1',
-            } as Gene);
-            genes.push({
-                entrezGeneId: 1,
-                hugoGeneSymbol: 'GENE2',
-            } as Gene);
-            genes.push({
-                entrezGeneId: 2,
-                hugoGeneSymbol: 'GENE3',
-            } as Gene);
-
-            genePanels.push({
-                genePanelId: 'GENEPANEL1',
-                genes: [genes[0], genes[1]],
-            });
-
-            genePanels.push({
-                genePanelId: 'GENEPANEL2',
-                genes: [genes[0], genes[1]],
-            });
-
-            samples.push({
-                uniqueSampleKey: 'PATIENT1 SAMPLE1',
-            } as Sample);
-            samples.push({
-                uniqueSampleKey: 'PATIENT1 SAMPLE2',
-            } as Sample);
-            samples.push({
-                uniqueSampleKey: 'PATIENT2 SAMPLE1',
-            } as Sample);
-
-            patients.push({
-                uniquePatientKey: 'PATIENT1',
-            } as Patient);
-            patients.push({
-                uniquePatientKey: 'PATIENT2',
-            } as Patient);
-
-            genePanelDatum1 = {
-                uniqueSampleKey: 'PATIENT1 SAMPLE1',
-                uniquePatientKey: 'PATIENT1',
-                molecularProfileId: 'PROFILE',
-                genePanelId: 'GENEPANEL1',
-                profiled: true,
-            };
-
-            genePanelDatum2 = {
-                uniqueSampleKey: 'PATIENT2 SAMPLE1',
-                uniquePatientKey: 'PATIENT2',
-                molecularProfileId: 'PROFILE',
-                genePanelId: 'GENEPANEL2',
-                profiled: true,
-            };
-
-            wxsDatum1 = {
-                uniqueSampleKey: 'PATIENT1 SAMPLE2',
-                uniquePatientKey: 'PATIENT1',
-                molecularProfileId: 'PROFILE',
-                profiled: true,
-            };
-
-            nsDatum1 = {
-                entrezGeneId: 2,
-                molecularProfileId: 'PROFILE',
-                uniqueSampleKey: 'PATIENT1 SAMPLE1',
-                uniquePatientKey: 'PATIENT1',
-                profiled: false,
-            };
-
-            nsDatum2 = {
-                entrezGeneId: 2,
-                molecularProfileId: 'PROFILE',
-                uniqueSampleKey: 'PATIENT2 SAMPLE1',
-                uniquePatientKey: 'PATIENT2',
-                profiled: false,
-            };
-        });
-        it('computes the correct object with no input data', () => {
-            assert.deepEqual(
-                computeGenePanelInformation(
-                    [],
-                    genePanels,
-                    samples,
-                    patients,
-                    genes
-                ),
-                {
-                    samples: {
-                        'PATIENT1 SAMPLE1': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT1 SAMPLE2': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT2 SAMPLE1': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                    },
-                    patients: {
-                        PATIENT1: {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        PATIENT2: {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                    },
-                }
-            );
-        });
-        it('computes the correct object with gene panel data', () => {
-            assert.deepEqual(
-                computeGenePanelInformation(
-                    [genePanelDatum1, genePanelDatum2] as GenePanelData[],
-                    genePanels,
-                    samples,
-                    patients,
-                    genes
-                ),
-                {
-                    samples: {
-                        'PATIENT1 SAMPLE1': {
-                            byGene: {
-                                GENE1: [genePanelDatum1],
-                                GENE2: [genePanelDatum1],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum1] },
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT1 SAMPLE2': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT2 SAMPLE1': {
-                            byGene: {
-                                GENE1: [genePanelDatum2],
-                                GENE2: [genePanelDatum2],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum2] },
-                            notProfiledAllGenes: [],
-                        },
-                    },
-                    patients: {
-                        PATIENT1: {
-                            byGene: {
-                                GENE1: [genePanelDatum1],
-                                GENE2: [genePanelDatum1],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum1] },
-                            notProfiledAllGenes: [],
-                        },
-                        PATIENT2: {
-                            byGene: {
-                                GENE1: [genePanelDatum2],
-                                GENE2: [genePanelDatum2],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum2] },
-                            notProfiledAllGenes: [],
-                        },
-                    },
-                }
-            );
-        });
-        it('computes the correct object with whole exome sequenced data', () => {
-            assert.deepEqual(
-                computeGenePanelInformation(
-                    [wxsDatum1] as GenePanelData[],
-                    genePanels,
-                    samples,
-                    patients,
-                    genes
-                ),
-                {
-                    samples: {
-                        'PATIENT1 SAMPLE1': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT1 SAMPLE2': {
-                            byGene: {},
-                            allGenes: [wxsDatum1],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT2 SAMPLE1': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                    },
-                    patients: {
-                        PATIENT1: {
-                            byGene: {},
-                            allGenes: [wxsDatum1],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        PATIENT2: {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                    },
-                }
-            );
-        });
-        it('computes the correct object with not sequenced data', () => {
-            assert.deepEqual(
-                computeGenePanelInformation(
-                    [nsDatum1, nsDatum2] as GenePanelData[],
-                    genePanels,
-                    samples,
-                    patients,
-                    genes
-                ),
-                {
-                    samples: {
-                        'PATIENT1 SAMPLE1': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [nsDatum1],
-                        },
-                        'PATIENT1 SAMPLE2': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT2 SAMPLE1': {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [nsDatum2],
-                        },
-                    },
-                    patients: {
-                        PATIENT1: {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [nsDatum1],
-                        },
-                        PATIENT2: {
-                            byGene: {},
-                            allGenes: [],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [nsDatum2],
-                        },
-                    },
-                }
-            );
-        });
-        it('computes the correct object with gene panel data and whole exome sequenced data' /*and not sequenced data"*/, () => {
-            assert.deepEqual(
-                computeGenePanelInformation(
-                    [
-                        genePanelDatum1,
-                        genePanelDatum2,
-                        wxsDatum1,
-                        nsDatum1,
-                        nsDatum2,
-                    ] as GenePanelData[],
-                    genePanels,
-                    samples,
-                    patients,
-                    genes
-                ),
-                {
-                    samples: {
-                        'PATIENT1 SAMPLE1': {
-                            byGene: {
-                                GENE1: [genePanelDatum1],
-                                GENE2: [genePanelDatum1],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum1] },
-                            notProfiledAllGenes: [nsDatum1],
-                        },
-                        'PATIENT1 SAMPLE2': {
-                            byGene: {},
-                            allGenes: [wxsDatum1],
-                            notProfiledByGene: {},
-                            notProfiledAllGenes: [],
-                        },
-                        'PATIENT2 SAMPLE1': {
-                            byGene: {
-                                GENE1: [genePanelDatum2],
-                                GENE2: [genePanelDatum2],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum2] },
-                            notProfiledAllGenes: [nsDatum2],
-                        },
-                    },
-                    patients: {
-                        PATIENT1: {
-                            byGene: {
-                                GENE1: [genePanelDatum1],
-                                GENE2: [genePanelDatum1],
-                            },
-                            allGenes: [wxsDatum1],
-                            notProfiledByGene: { GENE3: [genePanelDatum1] },
-                            notProfiledAllGenes: [nsDatum1],
-                        },
-                        PATIENT2: {
-                            byGene: {
-                                GENE1: [genePanelDatum2],
-                                GENE2: [genePanelDatum2],
-                            },
-                            allGenes: [],
-                            notProfiledByGene: { GENE3: [genePanelDatum2] },
-                            notProfiledAllGenes: [nsDatum2],
-                        },
-                    },
-                }
-            );
         });
     });
